done with plots
